Fixed the design -- now UART works in hardware test
I think Vivado does not update when HDL changes if that HDL is inside
a block design -- instead, if an RTL module is included inside a block
design, it seems like a snapshot is taken of it which is really hard
to update. (See
https://support.xilinx.com/s/question/0D52E00006hpM5SSAU/synthesis-does-not-update?language=en_US.)

After rearranging the design so that the top level module is an HDL
module, and the block design is only used to wrap the clocking wizard,
everything seems to update normally when the HDL changes (because the
top level module is a pure HDL module). This is also a better design,
because it modularises the clock component which can be swapped out
depending on the tool used.