Discovered failing induction due to potential infinite stall
The proof is failing because it finds a trace with a pending pop
transaction, where the buffer is empty. When it finally completes, the
counter wraps. The issue is that the induction cannot tell that the
transaction should not begin while buffer is empty.

One way to solve this problem would be to add a bound on the length of
time a device can stall for (artificially, to fix the
induction). However, a better solution would be to try to make the
logic "local" enough that a transaction cannot be _occurring_ while
the buffer is empty.

Going to try that first.