Fixed the leds module (typo, caught using `default_nettype none)
Unfortunately it does not seem possible to use `default_nettype none
in systemverilog, because it causes problems for input ports declared
as input logic (they are missing a net type). It feels like there
should be a way round this, but I'll leave it for now (the
`default_nettype did enough to catch the error, then I removed it).