Synthesis runs fine after changing internal interface names
Confirmed that the design now runs correctly on hardware.

I'm not sure whether this is an issue with the previous interface, or
a Vivado bug, but it is probably best to steer clear of interface
expressions in modports for simplicity; e.g. for Wishbone, can
standardise on using the controller signal names for both the
controller and device.