Started working on formal modules to simulate dev/ctrl behaviour
Previously, the assumptions for a simulated wishbone controller or
device were stored in the interface, and enabled by a define. However,
this does not work in a design under test which has multiple wishbone
interfaces, because you only want to enable the assumptions for the
external-facing interfaces rather than all of them.